bug fixes
the SFINAE i'd set up for the move constructor was bugging up on types that didn't satisfy it (not sure if this is because it was SFINAE on a ctor or if SFINAE that already knows its type is ill-formed if the resulting decl is ill-formed...).
regardless, the move ctor is now replaced with a broader version:
new version is still noexcept, but works for any type, regardless of noexcept move ctor quality.
internally uses std::move_if_noexept to make sure we don't lose information.
the other object (ctor source) is only emptied if that succeeds.

added constexpr to a couple of methods, esp. the function pointer specialization.