updated contract system
the old FATE contract system was truly Ate The End.
this is why a move assignment operator was not supported - because then the previous function-like object wouldn't be executed At The End.

the new contract idea takes more from the name FATE than the acronym - the function shall be called exactly once - it is destiny (unless, of course the release() method is called to abort it).

an operator() is added to invoke the stored function explicitly, after which the fate instance becomes empty.
a move assignment operator is added that calls operator() first to ensure it handles any function it might still have, then transfers the other object contract to itself (thus emptying other).